improvements on authentication

+++ Advanced_Todo_Application/routers/Todos.py
@@ -5,6 +5,7 @@ from typing import Annotated
 from sqlalchemy.orm import Session
 from DbTodosTableModel import Todos
 from DatabaseConnection import local_session
+from .UserAuthentication import get_current_user
 from TodoRequest import TodoRequest
 
 apiRouter = APIRouter()
@@ -26,6 +27,7 @@ def obtain_database():
 
 
 database_dependency = Annotated[Session, Depends(obtain_database)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 # 'Depends' import from FastAPI represents the dependency injection.
@@ -34,8 +36,8 @@ database_dependency = Annotated[Session, Depends(obtain_database)]
 # created. This database session is injected to the below function
 # as an instance called 'db'.
 @apiRouter.get("/", status_code=status.HTTP_200_OK)
-async def read_todos(database: database_dependency):
-    return database.query(Todos).all()
+async def read_todos(user: user_dependency, database: database_dependency):
+    return database.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 # The below function gets the todos that have the
@@ -47,17 +49,27 @@ the request was successful. This default HTTP response is set with the status_co
 
 
 @apiRouter.get("/getTodo/{todo_id}", status_code=status.HTTP_200_OK)
-async def getTodoById(database: database_dependency, todo_id: int = Path(gt=0)):
+async def getTodoById(user: user_dependency,
+                      database: database_dependency,
+                      todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="User Authentication is Failed")
     todo_model = database.query(Todos).filter(
-        Todos.todo_id == todo_id).first()  # get the first record where todo ids match. Due to the uniqueness of todo ids, using first() will be fine here.
+        Todos.todo_id == todo_id).filter(
+        Todos.owner_id == user.get('id')).first()  # get the first record where todo ids match. Due to the uniqueness of todo ids, using first() will be fine here.
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo is not found")
     return todo_model
 
 
 @apiRouter.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(database: database_dependency, todoReq: TodoRequest):
-    todo_object_model = Todos(**todoReq.model_dump())
+async def create_todo(
+        user: user_dependency,
+        database: database_dependency,
+        todoReq: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail="User authentication is failed.")
+    todo_object_model = Todos(**todoReq.model_dump(), owner_id=user.get('id'))
     database.add(todo_object_model)
     database.commit()  # this makes the above addition operation automatically happen.
 
@@ -67,9 +79,11 @@ async def create_todo(database: database_dependency, todoReq: TodoRequest):
 
 # In the below HTTP PUT request function, the todo_is is passed as a query parameter.
 @apiRouter.put("/todo/", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(database: database_dependency, todo_req: TodoRequest,
+async def update_todo(user: user_dependency,
+                      database: database_dependency, todo_req: TodoRequest,
                       todo_id: int = Query(description="The ID of the todo item to update", gt=0)):
-    todo_object_model = database.query(Todos).filter(Todos.todo_id == todo_id).first()
+    todo_object_model = database.query(Todos).filter(Todos.todo_id == todo_id).filter(Todos.owner_id == user.get('id')).first()
+    print(todo_object_model)
     if todo_object_model is None:
         raise HTTPException(status_code=404, detail='Todo cannot be found in the db table called todos')
 
@@ -80,6 +94,7 @@ async def update_todo(database: database_dependency, todo_req: TodoRequest,
 
     database.add(todo_object_model)
     database.commit()
+    print("Todo is successfully updated.")
 
 
 # HTTP PUT Request Method
@@ -90,13 +105,19 @@ async def update_todo(database: database_dependency, todo_req: TodoRequest,
 
 # In the below HTTP DELETE request function, the todo_is is passed as a path parameter.
 @apiRouter.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(database: database_dependency, todo_id: int = Path(gt=0)):
-    todo_object_model = database.query(Todos).filter(Todos.todo_id == todo_id).first()
+async def delete_todo(
+        user: user_dependency,
+        database: database_dependency,
+        todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='User authentication is failed.')
+    todo_object_model = database.query(Todos).filter(Todos.todo_id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_object_model is None:
         raise HTTPException(status_code=404, detail='Todo cannot be found in the db table called todos')
-    database.query(Todos).filter(Todos.todo_id == todo_id).delete()
-
+    database.query(Todos).filter(Todos.todo_id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     database.commit()
+    print("Todo is successfully deleted.")
 
 # HTTP DELETE Request Method
 ###################################################################################################
